# docs/api/phase1-auth.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ThunderForgeVTT Phase 1 Authentication and User Data API",
    "version": "v1",
    "description": "Public contract for local auth, OAuth completion/linking, session lifecycle, self-service data export, and permanent deletion."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin deployment"
    }
  ],
  "paths": {
...
    "/authentication/login": {
      "post": {
        "summary": "Sign in with local credentials",
        "responses": {
          "200": {
            "description": "Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or second factor required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthUser": {
        "type": "object",
        "required": ["id", "username", "email", "is_admin", "created_at", "updated_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "username": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "is_admin": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "AuthSessionResponse": {
        "type": "object",
        "required": ["status", "message", "session", "login_two_factor_challenge_id", "requires_email_verification"],
        "properties": {
          "status": { "type": "string" },
          "message": { "type": "string" },
          "login_two_factor_challenge_id": {
            "type": ["string", "null"],
            "format": "uuid"
          },
          "requires_email_verification": { "type": "boolean" },
          "session": {
            "type": ["object", "null"],
            "properties": {
              "authenticated": { "type": "boolean" },
              "session_expires_at": { "type": "string", "format": "date-time" },
              "user": { "$ref": "#/components/schemas/AuthUser" }
            }
          }
        }
      },
...
    }
  }
}